fix: remove redundant files and fix type name
