+++ app/frontend/src/App.tsx
@@ -1,18 +1,22 @@
 import React, { useEffect } from 'react';
-import WorkspaceView from './components/views/workspace-view/WorkspaceView';
 import WorkspaceForm from './components/views/workspace-form/WorkSpaceForm';
 import Container from 'components/layouts/base/Container';
 import Content from 'components/layouts/base/Content';
 import ColorVariables from './styles/Colors';
 import GlobalStyles from './styles/Global';
 import ProjectForm from 'components/views/project-form/ProjectForm';
-import { Navigate, Route, Routes } from 'react-router-dom';
+import { Link, Navigate, Route, Routes } from 'react-router-dom';
 import ProjectListView from 'components/views/project-list-view/ProjectListView';
 import ProjectRootView from 'components/views/project-root-view/ProjectRootView';
 import TaskPopup from 'components/views/project-list-view/components/TaskPopup';
 import ProjectBoardView from 'components/views/project-board-view/ProjectBoardView';
 import { useStore } from 'effector-react';
 import { $workspace, getWorkspaceFx } from 'store/workspaces';
+import SidebarHeader from './components/layouts/base/SidebarHeader';
+import ProjectList from './components/UI/project-list/ProjectList';
+import Sidebar from './components/layouts/base/Sidebar';
+import styled from 'styled-components';
+import Button from './components/UI/button/Button';
 
 /**
  * Makes the main page
@@ -24,13 +28,20 @@ function App(): React.ReactElement {
 
   useEffect(() => {
     getWorkspaceFx({});
-  }, []);
+  }, [ workspace.name ]);
 
   return (
     <Container>
       <ColorVariables/>
       <GlobalStyles/>
-      <WorkspaceView/>
+      <Sidebar>
+        <SidebarHeader sidebarTitle={ workspace.name !== null ? workspace.name : '' }/>
+        <ProjectList workspaceId={ workspace._id !== null ? workspace._id : '' }>
+          <StyledLink to="/projects/new">
+            <StyledButton icon="plus">Add new project</StyledButton>
+          </StyledLink>
+        </ProjectList>
+      </Sidebar>
       <Routes>
         <Route path="/" element={ <Content/> }>
           <Route path="workspace/edit" element={ <WorkspaceForm/> }/>
@@ -55,5 +66,19 @@ function App(): React.ReactElement {
     </Container>
   );
 }
+
+/**
+ * Link component with overriden styles
+ */
+const StyledLink = styled(Link)`
+  text-decoration: unset;
+`;
+
+/**
+ * Button component which takes full width
+ */
+const StyledButton = styled(Button)`
+  width: 100%;
+`;
 
 export default App;
